backend/requirements.txt
fastapi[standard]
sqlalchemy
asyncmy
greenlet
python-dotenv==1.0.1
cryptography==43.0.1
bcrypt==4.2.0
PyJWT==2.9.0